Create basic datalayer stub w/ D.I. using dagger2
Dagger 2 is love. Dagger 2 is life. This will be our basic application
data-layer in which we will have a `persistant` datalayer along with a
`network` datalayer which will both implement a singular interface as to
allow our view layer to be neglegent to where the data is being served
from.