Complete dagger2 impl. setup stubbable fragments.
This is the finalzied dagger2 setup for dependency injection. This
allows us to just have a more modularized project setup as well as
abstracting dependencies into proper scopes.

The fragment view stub is a usability feature as to enable me to
implement core functions at a time instead of implementing eveything at
once :-)